Slight update to extended Can directive
